Move vendor libraries to /vendors folder and exclude from LGTM
