+++ src/static/js/vendors/jquery.js
