+++ src/templates/admin/index.html
@@ -4,10 +4,10 @@
     <title data-l10n-id="admin.page-title">Admin Dashboard - Etherpad</title>
     <meta name="viewport" content="width=device-width">
     <link rel="stylesheet" href="../static/css/admin.css">
-    <script src="../static/js/jquery.js"></script>
+    <script src="../static/js/vendors/jquery.js"></script>
     <script src="../socket.io/socket.io.js"></script>
     <link rel="localizations" type="application/l10n+json" href="../locales.json" />
-    <script src="../static/js/html10n.js"></script>
+    <script src="../static/js/vendors/html10n.js"></script>
     <script src="../static/js/l10n.js"></script>
   </head>
   <body>
