+++ src/templates/admin/plugins-info.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width">
     <link rel="stylesheet" href="../../static/css/admin.css">
     <link rel="localizations" type="application/l10n+json" href="../../locales.json" />
-    <script src="../../static/js/html10n.js"></script>
+    <script src="../../static/js/vendors/html10n.js"></script>
     <script src="../../static/js/l10n.js"></script>
   </head>
   <body>
