+++ src/templates/javascript.html
@@ -9,24 +9,24 @@
     <body>
         <table id="jslicense-labels1">
             <tr>
-                <td><a href="/static/js/jquery-2.1.1.min.js">jquery-2.1.1.min.js</a></td>
+                <td><a href="/static/js/vendors/jquery-3.0.0.min.js">jquery-3.0.1.min.js</a></td>
                 <td><a href="http://www.jclark.com/xml/copying.txt">Expat</a></td>
-                <td><a href="/static/js/jquery.js">jquery.js</a></td>
+                <td><a href="/static/js/vendors/jquery.js">jquery.js</a></td>
             </tr>
             <tr>
-                <td><a href="/static/js/html10n.js">html10n.js</a></td>
+                <td><a href="/static/js/vendors/html10n.js">html10n.js</a></td>
                 <td><a href="http://www.jclark.com/xml/copying.txt">Expat</a></td>
-                <td><a href="/static/js/html10n.js">html10n.js</a></td>
+                <td><a href="/static/js/vendors/html10n.js">html10n.js</a></td>
             </tr>
             <tr>
-                <td><a href="/static/js/l10n.js">l10n.js</a></td>
+                <td><a href="/static/js/vendors/l10n.js">l10n.js</a></td>
                 <td><a href="http://www.apache.org/licenses/LICENSE-2.0">Apache-2.0-only</a></td>
-                <td><a href="/static/js/l10n.js">l10n.js</a></td>
+                <td><a href="/static/js/vendors/l10n.js">l10n.js</a></td>
             </tr>
             <tr>
-                <td><a href="/static/js/socket.io.js">socket.io.js</a></td>
+                <td><a href="/static/js/vendors/socket.io.js">socket.io.js</a></td>
                 <td><a href="http://www.jclark.com/xml/copying.txt">Expat</a></td>
-                <td><a href="/static/js/socket.io.js">socket.io.js</a></td>
+                <td><a href="/static/js/vendors/socket.io.js">socket.io.js</a></td>
             </tr>
             <tr>
                 <td><a href="/static/js/require-kernel.js">require-kernel.js</a></td>
