+++ src/templates/timeslider.html
@@ -45,7 +45,7 @@
 
   <link rel="localizations" type="application/l10n+json" href="../../locales.json" />
   <% e.begin_block("timesliderScripts"); %>
-  <script type="text/javascript" src="../../static/js/html10n.js?v=<%=settings.randomVersionString%>"></script>
+  <script type="text/javascript" src="../../static/js/vendors/html10n.js?v=<%=settings.randomVersionString%>"></script>
   <script type="text/javascript" src="../../static/js/l10n.js?v=<%=settings.randomVersionString%>"></script>
   <% e.end_block(); %>
 </head>
@@ -278,7 +278,7 @@
     require.setGlobalKeyPath("require");
 
     $ = jQuery = require('ep_etherpad-lite/static/js/rjquery').jQuery; // Expose jQuery #HACK
-    browser = require('ep_etherpad-lite/static/js/browser');
+    browser = require('ep_etherpad-lite/static/js/vendors/browser');
 
     var plugins = require('ep_etherpad-lite/static/js/pluginfw/client_plugins');
     var socket = require('ep_etherpad-lite/static/js/timeslider').socket;
