+++ src/tests/frontend/helper.js
@@ -7,7 +7,7 @@ const helper = {};
   const jsLibraries = {};
 
   helper.init = (cb) => {
-    $.get('/static/js/jquery.js').done((code) => {
+    $.get('/static/js/vendors/jquery.js').done((code) => {
       // make sure we don't override existing jquery
       jsLibraries.jquery = `if(typeof $ === 'undefined') {\n${code}\n}`;
